NTFS: Changed NTFS version fields to byte to match VolumeInformation attribute record

+++ DiskAccessLibrary/FileSystems/NTFS/FileRecord/FileRecord.cs
@@ -33,7 +33,7 @@ namespace DiskAccessLibrary.FileSystems.NTFS
         /// <remarks>
         /// https://blogs.technet.microsoft.com/askcore/2009/10/16/the-four-stages-of-ntfs-file-growth/
         /// </remarks>
-        public void UpdateSegments(int bytesPerFileRecordSegment, ushort minorNTFSVersion)
+        public void UpdateSegments(int bytesPerFileRecordSegment, byte minorNTFSVersion)
         {
             List<AttributeRecord> attributes = this.Attributes;
